Refactor?
